Fix numeric id query (#660)
* Fix numeric id queries using the new member query option.


* Add some helper methods to convert AnyList into typed arrays.


* Add helper to try to proxy any type.


* Split featureNumber function for performance improvement at other places.


* Fix that isInstance of members accept normal lists for TAG_MAP_FROM_ARRAY


* Add missing TAGLIST_CONTAINS, fix any type can be provided to detect op type, optimized casts


* Improve documentation of Tag queries, clarify that they operate on value, not key.


* Remove catching of exceptions, leaving this to the caller, needed for debugging and stack-tracing.


* Fix convertValue to work with all values, rename childType to componentType, add kotlin klass of types.


* Log all errors with stack-trace for better debugging, the error response does not contain stack trace!


* Use json_array_elements_text and add explicit cast, jsonb is slower here, because this is a one-time call.


* Improve array handling, whenever possible use native array without converting, improve type detection.


* Fix tag-map queries to use index and fix that queries on `id` work for numeric identifiers.


* Add tests for tag-map.


* Add fixes for tag-map in PgRows.


* Improve documentation


* Move toArray helpers from AnyList into ListProxy so they are available for all lists.


* Rename dict to book, minor performance improvement to avoid data copy for strings section in books.


* Improve static fromArray helper.


* Embed helpers only used by Tuple into Tuple to be improved further in the future.


* Remove TAG_MAP_FROM_ARRAY


* First part of replacing Int64 with Long


* Replace Int64 with Long


* Improve PgRows to supports arrays and consistent resize, fix failing tests.


* Fix that the name of an operation can be deducted from the instance type.


* Fix brackets in query builder and where builder.


* Fix activity log, do not mix old and new query style. Fix TagList test, default is normalized.


* Fix minor issue left from rebase and Int64 to Long switch.


* Added check that readFeatures.getFeatureIds() is empty


* Fix addTag call in constructor to use skipNormalize.


* Modify coerceSpatial in Tuple to make the execution path clearer.


* Fix that when an operation is a no-op, AND and OR to not linger.


* Fix tests that faulty expect not normalized tag.


* Fix that only sourceID_ has special treatment.


* Fix bug in value iteration, added tests for sourceID_ and at tags.


* Restore tests to original casing.


* Make tag normalization optional, defaulting to none, adding normalize helper.


* Improve constructors of TagMap, update documentation.


* Remove renamed skipNormalize parameter, is now normalize.


* Improve normalizer, use last equals for split instead of first, added join method. Improve tests to match fixes.


* Fix error message of joinTag, fix test to expect more error messages.


---------

+++ here-naksha-app-service/src/jvmTest/java/com/here/naksha/app/common/FeatureMetadata.java
@@ -19,8 +19,8 @@ public record FeatureMetadata(String featureId, String uuid, long createdAt, lon
     return new FeatureMetadata(
         feature.getId(),
         xyzNamespace.getUuid(),
-        xyzNamespace.getCreatedAt().toLong(),
-        xyzNamespace.getUpdatedAt().toLong()
+        xyzNamespace.getCreatedAt(),
+        xyzNamespace.getUpdatedAt()
     );
   }
